Adding support for explicit element names and decoding (see DERDecoderFunctionalSpec.scala) * explicit naming is meant to aid with diagnosing cause of decoding error, but does not work well with decoding, so might change * decoding add the Value class (Constant/Placeholder) but (for now) only Constant works with encoding. Considering if Placeholder will fit in.

+++ src/test/scala/asn1/encoding/BitStringTagSpec.scala
@@ -12,7 +12,7 @@ class BitStringTagSpec extends FlatSpec with TableDrivenPropertyChecks {
   val encodingValue =
     Table(
       ("input", "expected"),
-      (BitString(Containing(Integer(1))), "030400020101"),
+      (BitString(Containing(Integer(BigInt(1)))), "030400020101"),
       (BitString(Literal(Seq(0x7f, 0x77), 0)), "0303007F77"),
       (BitString(Literal(Seq(0x7f, 0x70), 4)), "0303047F70")
       )
